% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_qtl_marker_fragment}
\alias{get_qtl_marker_fragment}
\title{This is a function that can return the marker fragment of QTL.}
\usage{
get_qtl_marker_fragment(qtlname = "qISO", location = c(5, 317.01, 319.09))
}
\arguments{
\item{qtlname}{is the name of QTL. you can define the name.}

\item{location}{is a number vactor that contein three elements.
the first is the chromosome number, the second and third is the
start and end of cM distance.}
}
\description{
This is a function that can return the marker fragment of QTL.
}
