% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_traits_annotation}
\alias{get_traits_annotation}
\title{Find the annotation by trait given.}
\usage{
get_traits_annotation(trait = "isoflavone")
}
\arguments{
\item{trait}{is a soybean trait.}
}
\description{
Find the annotation by trait given.
}
\examples{
get_traits_annotation("isoflavone")
}
